Make nogravity work on 64 bit machines
Was a lot harder then I initially thought dough, see the patch.
Especially the fact that the V3KKEY struct needs to be at the same offset
in several structs as those are crammed into a union, and the code expects
to be able to access this struct without taking into account which struct
actually sits inside the union.

Pointers on x86_64 will get 64 bits aligned, this takes this into account.

I've tested both with opengl and without and with openal and sdlmixer, all
work fine. This was all done on a clean x86_64 F-8 install.

Fedora patch: nogravity--64-bit.patch
First posted as https://bugzilla.redhat.com/show_bug.cgi?id=366841
attachment 263141 (2007-11-18); this is the final version, attachment 265081.

+++ rlx32/src/gx_csp.c
@@ -443,7 +443,7 @@ _RLXEXPORTFUNC void CSP_Resize(GXSPRITE *sp, int lx, int ly, int bpp)
 {
     u_int32_t sz = lx*ly, 
     sz2 = sz*bpp, 
-    ok = (((u_int32_t)MM_heap.PreviousAddress==(u_int32_t)sp->data)&&(MM_heap.active));
+    ok = (((uintptr_t)MM_heap.PreviousAddress==(uintptr_t)sp->data)&&(MM_heap.active));
     u_int8_t *tmp;
     tmp = (u_int8_t *)MM_heap.malloc(sz2);
     IMG_stretch(sp->data, tmp, sp->LY, ly, sp->LX, lx, bpp);
